requirements/local.txt
-r base.txt

django-extensions==3.2.3
pytest
pytest-django
local
send2trash
Werkzeug==3.0.3
pyOpenSSL==24.1.0
black==24.4.2
isort==5.13.2
factory-boy==3.2.1
freezegun==1.2.2
